deployment3/requirements.txt
streamlit
tensorflow-cpu
scikit-learn
pillow
numpy